AI_Agents_Workshop/App/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.30.0
python-multipart==0.0.6
psycopg2-binary==2.9.6  # For PostgreSQL support
python-dotenv==1.0.0 